Added in the initialization code for the blob detector app.
Currently, it just initializes the AXI DMA device, mounts the SD card as
a FAT filesystem, and reads a file on the SD card. One thorny issue is
that fact that `xilffs` does not come with LFN (long filename) support.
This each component of a file path (including file extensions) is
limited to 8 characters, and the extension is limited to 3 characters.